+++ src/shared/time_op.c
@@ -82,6 +82,21 @@ void gettime(struct timespec * ts) {
 
 #endif
 
+char *w_get_timestamp(time_t time) {
+    struct tm localtm;
+    char *timestamp;
+
+    localtime_r(&time, &localtm);
+
+    os_calloc(TIME_LENGTH, sizeof(char), timestamp);
+
+    snprintf(timestamp, TIME_LENGTH, "%d/%02d/%02d %02d:%02d:%02d",
+            localtm.tm_year + 1900, localtm.tm_mon + 1,
+            localtm.tm_mday, localtm.tm_hour, localtm.tm_min, localtm.tm_sec);
+
+    return timestamp;
+}
+
 // Compute time substraction "a - b"
 
 void time_sub(struct timespec * a, const struct timespec * b) {