+++ src/wazuh_db/schema_agents.sql
@@ -100,6 +100,7 @@ CREATE TABLE IF NOT EXISTS sys_osinfo (
     sysname TEXT,
     release TEXT,
     version TEXT,
+    os_release TEXT,
     PRIMARY KEY (scan_id, os_name)
 );
 
@@ -151,10 +152,21 @@ CREATE TABLE IF NOT EXISTS sys_programs (
     description TEXT,
     location TEXT,
     triaged INTEGER(1),
+    cpe TEXT,
+    msu_name TEXT,
     PRIMARY KEY (scan_id, name, version, architecture)
 );
 
 CREATE INDEX IF NOT EXISTS programs_id ON sys_programs (scan_id);
+
+CREATE TABLE IF NOT EXISTS sys_hotfixes (
+    scan_id INTEGER,
+    scan_time TEXT,
+    hotfix TEXT,
+    PRIMARY KEY (scan_id, scan_time, hotfix)
+);
+
+CREATE INDEX IF NOT EXISTS hotfix_id ON sys_hotfixes (scan_id);
 
 CREATE TABLE IF NOT EXISTS sys_processes (
     scan_id INTEGER,
@@ -285,6 +297,15 @@ CREATE TABLE IF NOT EXISTS sca_check_compliance (
 );
 
 CREATE INDEX IF NOT EXISTS comp_id_check_index ON sca_check_compliance (id_check);
+
+CREATE TABLE IF NOT EXISTS vuln_metadata (
+    LAST_SCAN INTEGER,
+    WAZUH_VERSION TEXT
+);
+INSERT INTO vuln_metadata (LAST_SCAN, WAZUH_VERSION)
+    SELECT '0', '0' WHERE NOT EXISTS (
+        SELECT * FROM vuln_metadata
+    );
 
 CREATE TABLE IF NOT EXISTS sync_info (
     component TEXT PRIMARY KEY,