+++ src/wazuh_db/wdb.h
@@ -60,7 +60,9 @@ typedef enum wdb_stmt {
     WDB_STMT_PROGRAM_DEL,
     WDB_STMT_PROGRAM_UPD,
     WDB_STMT_HWINFO_INSERT,
+    WDB_STMT_HOTFIX_INSERT,
     WDB_STMT_HWINFO_DEL,
+    WDB_STMT_HOTFIX_DEL,
     WDB_STMT_PORT_INSERT,
     WDB_STMT_PORT_DEL,
     WDB_STMT_PROC_INSERT,
@@ -423,6 +425,9 @@ int wdb_netinfo_save(wdb_t * wdb, const char * scan_id, const char * scan_time, 
 // Delete Network info from DB.
 int wdb_netinfo_delete(wdb_t * wdb, const char * scan_id);
 
+// Delete Hotfix info from DB.
+int wdb_hotfix_delete(wdb_t * wdb, const char * scan_id);
+
 // Insert IPv4/IPv6 protocol info tuple. Return 0 on success or -1 on error.
 int wdb_netproto_insert(wdb_t * wdb, const char * scan_id, const char * iface,  int type, const char * gateway, const char * dhcp, int metric);
 
@@ -436,10 +441,10 @@ int wdb_netaddr_insert(wdb_t * wdb, const char * scan_id, const char * iface, in
 int wdb_netaddr_save(wdb_t * wdb, const char * scan_id, const char * iface, int proto, const char * address, const char * netmask, const char * broadcast);
 
 // Insert OS info tuple. Return 0 on success or -1 on error.
-int wdb_osinfo_insert(wdb_t * wdb, const char * scan_id, const char * scan_time, const char * hostname, const char * architecture, const char * os_name, const char * os_version, const char * os_codename, const char * os_major, const char * os_minor, const char * os_build, const char * os_platform, const char * sysname, const char * release, const char * version);
+int wdb_osinfo_insert(wdb_t * wdb, const char * scan_id, const char * scan_time, const char * hostname, const char * architecture, const char * os_name, const char * os_version, const char * os_codename, const char * os_major, const char * os_minor, const char * os_build, const char * os_platform, const char * sysname, const char * release, const char * version, const char * os_release);
 
 // Save OS info into DB.
-int wdb_osinfo_save(wdb_t * wdb, const char * scan_id, const char * scan_time, const char * hostname, const char * architecture, const char * os_name, const char * os_version, const char * os_codename, const char * os_major, const char * os_minor, const char * os_build, const char * os_platform, const char * sysname, const char * release, const char * version);
+int wdb_osinfo_save(wdb_t * wdb, const char * scan_id, const char * scan_time, const char * hostname, const char * architecture, const char * os_name, const char * os_version, const char * os_codename, const char * os_major, const char * os_minor, const char * os_build, const char * os_platform, const char * sysname, const char * release, const char * version, const char * os_release);
 
 // Insert HW info tuple. Return 0 on success or -1 on error.
 int wdb_hardware_insert(wdb_t * wdb, const char * scan_id, const char * scan_time, const char * serial, const char * cpu_name, int cpu_cores, const char * cpu_mhz, uint64_t ram_total, uint64_t ram_free, int ram_usage);
@@ -452,6 +457,12 @@ int wdb_package_insert(wdb_t * wdb, const char * scan_id, const char * scan_time
 
 // Save Packages info into DB.
 int wdb_package_save(wdb_t * wdb, const char * scan_id, const char * scan_time, const char * format, const char * name, const char * priority, const char * section, long size, const char * vendor, const char * install_time, const char * version, const char * architecture, const char * multiarch, const char * source, const char * description, const char * location);
+
+// Insert hotfix info tuple. Return 0 on success or -1 on error.
+int wdb_hotfix_insert(wdb_t * wdb, const char * scan_id, const char * scan_time, const char *hotfix);
+
+// Save Hotfixes info into DB.
+int wdb_hotfix_save(wdb_t * wdb, const char * scan_id, const char * scan_time, const char *hotfix);
 
 // Update the new Package info with the previous scan.
 int wdb_package_update(wdb_t * wdb, const char * scan_id);
@@ -531,6 +542,8 @@ int wdb_parse_hardware(wdb_t * wdb, char * input, char * output);
 
 int wdb_parse_packages(wdb_t * wdb, char * input, char * output);
 
+int wdb_parse_hotfixes(wdb_t * wdb, char * input, char * output);
+
 int wdb_parse_ports(wdb_t * wdb, char * input, char * output);
 
 int wdb_parse_processes(wdb_t * wdb, char * input, char * output);
@@ -588,5 +601,8 @@ int wdbi_query_checksum(wdb_t * wdb, wdb_component_t component, const char * com
  * @retval -1 On error.
  */
 int wdbi_query_clear(wdb_t * wdb, wdb_component_t component, const char * payload);
+
+// Finalize a statement securely
+#define wdb_finalize(x) { if (x) { sqlite3_finalize(x); x = NULL; } }
 
 #endif